requirements.txt
fastapi
pydantic
SQLAlchemy
uvicorn
psycopg2-binary
typing_extensions
python-dotenv
apscheduler
httpx
pyngrok